requirements.txt
requests
moviepy
pytube3
youtube_transcript_api
SpeechRecognition
pydub
deep-translator
transformers
huggingsound 
librosa
nltk
rouge
matplotlib
seaborn
scikit-learn
torch
validators
Flask
Flask-Cors
numpy
regex
tqdm
urllib3
python-dotenv
os